requirements.txt
agno>=0.7.0
groq>=0.8.0
streamlit>=1.28.0
pyodbc>=4.0.39
pymssql>=2.2.0
SQLAlchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio>=3.4.3